rearranging deck chairs
