+++ py/gpu_specter/io.py
@@ -3,6 +3,20 @@ import numpy as np
 from astropy.table import Table
 import fitsio
 
+def native_endian(data):
+    """Temporary function, sourced from desispec.io
+    Convert numpy array data to native endianness if needed.
+    Returns new array if endianness is swapped, otherwise returns input data
+    Context:
+    By default, FITS data from astropy.io.fits.getdata() are not Intel
+    native endianness and scipy 0.14 sparse matrices have a bug with
+    non-native endian data.
+    """
+    if data.dtype.isnative:
+        return data
+    else:
+        return data.byteswap().newbyteorder()
+
 def read_psf(filename):
     """
     Read GaussHermite PSF data from input filename
@@ -33,47 +47,3 @@ def read_psf(filename):
     
     return psf
 
-#- subset of desiutil.log.get_logger, to avoid desiutil dependency
-_loggers = dict()
-def get_logger(level=None):
-
-    if level is None:
-        level = os.getenv('DESI_LOGLEVEL', 'INFO')
-
-    level = level.upper()
-    if level == 'DEBUG':
-        loglevel = logging.DEBUG
-    elif level == 'INFO':
-        loglevel = logging.INFO
-    elif level == 'WARN' or level == 'WARNING':
-        loglevel = logging.WARNING
-    elif level == 'ERROR':
-        loglevel = logging.ERROR
-    elif level == 'FATAL' or level == 'CRITICAL':
-        loglevel = logging.CRITICAL
-    else:
-        raise ValueError('Unknown log level {}; should be DEBUG/INFO/WARNING/ERROR/CRITICAL'.format(level))
-
-    if level not in _loggers:
-        logger = logging.getLogger('desimeter.'+level)
-        logger.setLevel(loglevel)
-
-        #- handler and formatter code adapted from
-        #- https://docs.python.org/3/howto/logging.html#configuring-logging
-
-        # create console handler and set level to debug
-        ch = logging.StreamHandler()
-        ch.setLevel(loglevel)
-
-        # create formatter
-        formatter = logging.Formatter('%(levelname)s:%(filename)s:%(lineno)s:%(funcName)s:%(message)s')
-
-        # add formatter to ch
-        ch.setFormatter(formatter)
-
-        # add ch to logger
-        logger.addHandler(ch)
-
-        _loggers[level] = logger
-
-    return _loggers[level]
